review_branch: Introduce a DEFAULT_BASE variable
Instead of assuming it way down in the code, it is better to set it up
front at the top itself. This would make it easier to transition to a
proper fix later on. Moreover, if a user needs to use a different base
before a proper fix gets deployed, he/she can easily edit the
DEFAULT_BASE variable itself instead of hunting for the location in the
code.